devops_api/requirements.txt
aioredis==1.3.1
amqp==2.6.1
asgiref==3.4.1
async-timeout==4.0.2
attrs==22.2.0
autobahn==21.2.1
Automat==22.10.0
bcrypt==4.0.1
billiard==3.6.4.0
celery==4.4.7
certifi==2022.12.7
cffi==1.15.1
channels==2.3.1
channels-redis==2.4.1
charset-normalizer==2.0.12
constantly==15.1.0
cryptography==39.0.2
daphne==2.5.0
Django==3.2.13
django-celery-beat==2.0.0
django-celery-results==2.0.0
django-cors-headers==3.10.1
django-redis==5.2.0
django-simpleui==2023.3.1
django-timezone-field==4.2.3
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
et-xmlfile==1.1.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.4
importlib-metadata==4.8.3
incremental==22.10.0
kombu==4.6.11
msgpack==0.6.2
multi-key-dict==2.0.3
mysqlclient==2.1.1
openpyxl==3.1.1
packaging==21.3
paramiko==3.0.0
pbr==5.11.1
Pillow==8.4.0
psutil==5.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycrypto==2.6.1
PyJWT==1.7.1
PyNaCl==1.5.0
pyOpenSSL==23.0.0
pyparsing==3.0.9
python-crontab==2.7.1
python-dateutil==2.8.2
python-gitlab==2.10.1
python-jenkins==1.7.0
pytz==2022.7.1
redis==4.3.5
requests==2.27.1
requests-toolbelt==0.10.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.1.1
urllib3==1.26.15
vine==1.3.0
zipp==3.6.0
zope.interface==5.5.2
